test PostRepository

+++ src/main/java/com/Deadline/BackEnd/Backend/exception/UerNotFoundExcetion.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class UerNotFoundExcetion extends  RuntimeException{
-    public UerNotFoundExcetion(Integer id){
+    public UerNotFoundExcetion(Long id){
         super("User not found with"+id);
     }
